# docs/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "LinkShorty API"
  description: "API для управления короткими ссылками"
host: localhost:8081
basePath: /
schemes:
  - http
securityDefinitions:
  BearerAuth:
    type: apiKey
    in: header
    name: Authorization
paths:
  /auth/login:
    post:
      summary: "Вход пользователя"
      description: "Аутентифицирует пользователя и возвращает JWT токен"
      tags:
        - auth
      consumes:
        - application/json
      parameters:
        - name: request
          in: body
          required: true
          schema:
            $ref: '#/definitions/auth.LoginRequest'
      responses:
        "200":
          description: "Успешный вход"
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: "Неверные учетные данные"
        "401":
          description: "Неверные учетные данные"
        "500":
          description: "Внутренняя ошибка сервера"
  /auth/register:
    post:
      summary: "Регистрация нового пользователя"
      description: "Регистрирует нового пользователя в системе"
      tags:
        - auth
      consumes:
        - application/json
      parameters:
        - name: request
          in: body
          required: true
          schema:
            $ref: '#/definitions/auth.RegisterRequest'
      responses:
        "201":
          description: "Успешная регистрация"
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: "Неверный формат запроса"
        "500":
          description: "Внутренняя ошибка сервера"
...
definitions:
...
  auth.LoginRequest:
    type: object
    properties:
      email:
        type: string
        example: user@example.com
      password:
        type: string
        example: strongPassword123
  auth.LoginResponse:
    type: object
    properties:
      token:
        type: string
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
  auth.RegisterRequest:
    type: object
    properties:
      email:
        type: string
        example: user@example.com
      name:
        type: string
        example: John Doe
      password:
        type: string
        example: strongPassword123
  auth.RegisterResponse:
    type: object
    properties:
      token:
        type: string
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...